Add NKG init-file including realigned velocity grid.
The NKG init-file defines transformations going to and from
the common nordic NKG reference frame using the Scandinavian
uplift model. The model is represented by two grids specially
made for the purpose. The grids are created with the
realign_nkgrf03vel.py script, which realigns the original
uplift model to the NKG_ETRF00 frame as described in the paper
by Häkli et al (2016).