Remove null grids from deformation operations
The null grids were added to avoid crashes when trying to transform
coordinates outside the deformation grids. This unfortunately made
the deformation operation use the null grid instead of the actual
deformation model grids. The bug that caused PROJ to crash has since
been fixed and the null grids can be removed from the +xy_grids and
+z_grids parameters.

Closes #9